basira/pyproject.toml
[project]
name = "basira"
version = "0.1.0"
description = "Basira — an independent geopolitical research API that monitors think tanks, official sources, multilaterals, and long-form analysis, and produces structured analytical scaffolding via Claude."
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database + migrations
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "sqlite-vec>=0.1.6",
    # Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    # Ingestion
    "feedparser>=6.0.11",
    "httpx>=0.28.0",
    "selectolax>=0.3.25",
    "trafilatura>=1.12.0",
    "pymupdf>=1.24.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    # AI
    "anthropic>=0.39.0",
    "voyageai>=0.3.0",
    # Supabase (headless worker path — writes to Postgres via PostgREST)
    "supabase>=2.9.0",
    # Scheduler
    "apscheduler>=3.10.4",
    # Templates
    "jinja2>=3.1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "UP"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
